+++ package.json
@@ -28,8 +28,8 @@
     "react-query": "^3.39.2",
     "react-slick": "^0.29.0",
     "recoil": "^0.7.6",
-    "slick-carousel": "^1.8.1",
     "recoil-persist": "^4.2.0",
+    "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.6",
     "styled-reset": "^4.4.4",
     "typescript": "4.9.4"
