+++ src/constants/organizerOnboarding/organizerSliderItems.ts
@@ -1,0 +1,29 @@
+import { imgOnbarding1, imgOnbarding2, imgOnbarding3, imgOnbarding4, imgOnbarding5 } from '@src/assets/images';
+
+export const ORGANIZER_SLIDER_ITEMS = [
+  {
+    id: 1,
+    text: '팀워크를 높이고 싶으신가요?\n저 티타미가 도와드릴게요!',
+    image: imgOnbarding1,
+  },
+  {
+    id: 2,
+    text: '초대장을 만들어\n팀원들에게 공유해보세요.',
+    image: imgOnbarding2,
+  },
+  {
+    id: 3,
+    text: "'티타미'와의 채팅을 통해\n개인의 감정을 털어놓고",
+    image: imgOnbarding3,
+  },
+  {
+    id: 4,
+    text: '개인 혹은 팀 리포트를 통해\n우리팀의 생각을 엿볼 수 있어요.',
+    image: imgOnbarding4,
+  },
+  {
+    id: 5,
+    text: '팀원들과 차한잔 마시면서\n가볍게 이야기 시작해보세요!',
+    image: imgOnbarding5,
+  },
+];
